Refactored all code into a new class called Flaccurate
There was a growing amount of global state which was being
initialised in functions at a distance, getting a bit messy.

The new class Flaccurate now holds all the appropriate state
as instance data with the methods acting upon it.

Make use of the previously implemented __main__: main() function to
wrap up the instantiation of the new Flaccurate class and invokation
of the process_all method which does the business.

Signed-off-by: Kayvan Javid <javid.kayvan@gmail.com>