Completed database integrity checksum updating
The previous commit which completed _valid_db() implemented
the reading of the checksum.
This commit completes the writing and on-going maintenance
of the checksum.

After the database is initially created it gets an md5 checksum
entry in a file named after it with an .md5 extension.  This value
is then maintained on any database modification - i.e. after every
insert.

Next time _valid_db() is called (on subsequent flaccurate
invokations), the value in this .md5 file is compared
against a freshly calc'd md5 to verify the db has not changed.

Signed-off-by: Kayvan Javid <javid.kayvan@gmail.com>