Rework plugin loading code motivated by binary distribution
Having tried many freezing applications, with varying degrees of
failure.  The common problem with them all seemed to be the dynamic
loading of modules from the plugins/ folder.  The implementation
always looked a bit fragile, regexp matching filenames in a static
folder, and calling import_module.

Have instead chosen to get __init__.py in the plugins/ folder to do
all the heavy lifting instead.  It enumerates all the plugins available,
(flac and mp3 at the time of writing) and imports them.  This means
using them only requires a simple: import flaccurate.plugins.

One benefit of the original file/folder traversal and loading was the
ability to explicitly know what plugins are available at the time
they are found on disk and loaded.  In the reworked implementation,
it is accomploshed by parsing the modules dictionary after they are
all imported via: sys.modules.keys() to find those where the name
starts: flaccurate.plugins.

Also of note in this commit:
 ^ Moved plugins/ into flaccurate/, providing a base namespace for
   all modules and packages of flaccurate.
 * Added additional command of: -c --selfcheck
   Simply initialises Flaccurate, its database, and loads all
   plugins.  Really useful to check runtime errors at startup.

Signed-off-by: Kayvan Javid <javid.kayvan@gmail.com>