Refactor database connect and validation code
Was connecting to the db in two separate places and duplicating
the error handling in both places.  Now the connect is encapsulated
in a single call, with the appopriate try/except semantics, which
includes a: raise RuntimeError on error.  This propogates back up
to the caller instantiating the Database object, which is the top
level flaccurate.py and errors out appropriately.

Also cleaned up the valid_db code with calls to the two separate
and distinct validation routines.  Now encapsulated in individual
private methods, namely:
_db_integrity_verified()
_db_checksum_verified()

Signed-off-by: Kayvan Javid <javid.kayvan@gmail.com>