Added --force implementation to database handling
Now there are several permutations.  If no database file is specified
the default behaviour stays the same, we use / creating one called
flaccurate.db.

If a database is specified (--database) and is not found, we raise a
Runtime exception instead of trying to guess what the user wants to do.
i.e. If you manually specify a database were you expecting us to find
it? If so, then erroring out is the right thing to do.

If a database is specified (--database) and is not found,
and force is specified (--force) we ignore not finding it and create
a new database of the name specified.  This is legitimate for a db
of a custom name on first run for example.

Signed-off-by: Kayvan Javid <javid.kayvan@gmail.com>